Move `ManufacturerDeviceV2` to Frient module, as other quirks will also use it
