+++ zhaquirks/develco/emi.py
@@ -3,42 +3,18 @@
 from typing import Final
 
 from zigpy.quirks import CustomCluster
-from zigpy.quirks.v2 import CustomDeviceV2, QuirkBuilder
+from zigpy.quirks.v2 import QuirkBuilder
 import zigpy.types as t
 from zigpy.zcl.clusters.smartenergy import Metering
 from zigpy.zcl.foundation import BaseAttributeDefs, ZCLAttributeDef
 
-
-class ManufacturerDevice(CustomDeviceV2):
-    """Custom device class used to remap cluster IDs in requests."""
-
-    async def request(
-        self,
-        *args,
-        **kwargs,
-    ):
-        """Remap cluster IDs for clusters that substitute for others."""
-        endpoint_id = kwargs["src_ep"]
-        cluster_id = kwargs["cluster"]
-
-        # ignore ZDO and narrow down to manufacturer specific clusters
-        if endpoint_id != 0 and cluster_id >= 0xFC00:
-            cluster = self.endpoints[endpoint_id].in_clusters.get(cluster_id)
-            substitution_cluster = getattr(cluster, "SUBSTITUTION_FOR", None)
-
-            if substitution_cluster:
-                kwargs["cluster"] = t.ClusterId(substitution_cluster)
-
-        return await super().request(
-            *args,
-            **kwargs,
-        )
+from zhaquirks.develco import ManufacturerDeviceV2
 
 
 class ManufacturerMetering(CustomCluster):
     """Fake manufacturer specific cluster for Metering manufacturer attributes."""
 
-    SUBSTITUTION_FOR: int = Metering.cluster_id
+    SUBSTITUTION_FOR = Metering.cluster_id  # requires ManufacturerDeviceV2
     cluster_id = 0xFD10
 
     class AttributeDefs(BaseAttributeDefs):
@@ -54,7 +30,7 @@ class ManufacturerMetering(CustomCluster):
 base_quirk = (
     QuirkBuilder()
     .replaces(ManufacturerMetering, endpoint_id=2)
-    .device_class(ManufacturerDevice)
+    .device_class(ManufacturerDeviceV2)
     .number(
         attribute_name=ManufacturerMetering.AttributeDefs.pulse_configuration.name,
         cluster_id=ManufacturerMetering.cluster_id,
